+++ studio_version.py
@@ -1,2 +1,2 @@
 """Package release version."""
-VERSION = "0.3.3"
+VERSION = "0.3.4"